+++ package.json
@@ -9,14 +9,14 @@
     "deploy": "docusaurus deploy",
     "serve": "docusaurus serve",
     "clear": "docusaurus clear",
-    "test": "jest ./src/*",
+    "test": "jest ./src/* --passWithNoTests",
     "test:single": "jest --watch"
   },
   "dependencies": {
     "@docusaurus/core": "^2.0.0-beta.18",
     "@docusaurus/preset-classic": "^2.0.0-beta.18",
     "@govtechsg/tradetrust-ui-components": "^2.16.0",
-    "@govtechsg/tradetrust-utils": "^1.2.0",
+    "@govtechsg/tradetrust-utils": "^1.2.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
