--- src/theme/DocItem/index.test.tsx
@@ -1,8 +1,0 @@
-// import React from "react";
-// import { render, screen, fireEvent } from "@testing-library/react";
-
-describe("FeedbackWidget", () => {
-  it("wip", () => {
-    expect(true).toBe(true); // wip
-  });
-});
